Rename package ditto.internal.mvcc -> ditto.internal.record
