+++ ditto/src/main/java/eu/xenit/testing/ditto/internal/DefaultContentView.java
@@ -8,8 +8,8 @@ import eu.xenit.testing.ditto.api.model.Transaction;
 import eu.xenit.testing.ditto.api.data.ContentModel.Content;
 import eu.xenit.testing.ditto.api.model.QName;
 import eu.xenit.testing.ditto.internal.content.InternalContentData;
-import eu.xenit.testing.ditto.internal.mvcc.Cursor;
-import eu.xenit.testing.ditto.internal.mvcc.RecordLogEntry;
+import eu.xenit.testing.ditto.internal.record.Cursor;
+import eu.xenit.testing.ditto.internal.record.RecordLogEntry;
 import java.io.InputStream;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
