+++ ditto/src/main/java/eu/xenit/testing/ditto/internal/record/RecordLog.java
@@ -1,4 +1,4 @@
-package eu.xenit.testing.ditto.internal.mvcc;
+package eu.xenit.testing.ditto.internal.record;
 
 import java.util.Objects;
 import java.util.function.BinaryOperator;
