+++ ditto/src/main/java/eu/xenit/testing/ditto/internal/record/RecordLogEntry.java
@@ -1,4 +1,4 @@
-package eu.xenit.testing.ditto.internal.mvcc;
+package eu.xenit.testing.ditto.internal.record;
 
 import java.util.Objects;
 import java.util.UUID;
